refactor(db): extract SQLite UDF bindings behind a weak attach hook (Phase C.2a)
The per-runtime db.udf bindings (register/unregister + the sqlite3_value/
context marshalling) lived inline in mod_db.c, so the embedded runtime archive
libhull_feature-<rt>.a carried 17 undefined sqlite3_* references. A db-free app
whole-archiving that runtime bridge dragged them in, forcing the SQLite engine
even when unused: the last coupling keeping a SQLite-less base from producing a
SQLite-less app (docs/sqlite_feature.md, the Phase C limitation NOTE).

Split the udf region into a per-runtime bridge TU (runtime/{lua,js}/mod_db_udf.c)
behind a weak seam. mod_db.c now calls a weak hl_{lua,js}_db_attach_udf no-op;
the bridge provides the strong override that builds the udf sub-object. The JS
bridge reuses two mod_db.c helpers (js_call_handle, new_bound_subobject) via a
new internal header runtime/js/mod_db.h; the Lua bridge is self-contained.

Result: mod_db.o carries zero sqlite3_* references in both runtimes (nm-verified),
mod_db.c drops its sqlite3.h / db_sqlite.h / db_udf.h / wasm.h includes. This is
a pure refactor: the bridge object still ships in the runtime archive for now, so
behaviour and archive contents are unchanged. Phase C.2b moves it into its own
composed archive so the archive itself goes SQLite-free.

Verified: make test 60/60; db.udf works end to end (Lua exit 0; JS registers and
returns correct results). A pre-existing JS-only teardown assertion
(JS_FreeRuntime GC-leak on app.main + udf clean exit) reproduces on unmodified
main and is tracked separately.
